Move implementations to AbstractConvertService
There was no logic that needed to be specified outside an Abstract
layer.
